infound-data-collection/pyproject.toml
[project]
name = "infound-data-collection"
version = "1.0.0"
description = "INFound data collection services"
authors = [
    {name = "INFound"}
]
readme = "readme.md"

requires-python = ">=3.13,<4.0"
dependencies = [
    "pydantic (>=2.12.5)",
    "pydantic-settings (>=2.12.0)",
    "httpx (>=0.27.0)",
    "sqlalchemy (>=2.0.31)",
    "asyncmy (>=0.2.10)",
    "pyyaml (>=6.0.2)",
    "structlog (>=24.3.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "playwright (>=1.56.0,<2.0.0)",
    "concurrent-log-handler (>=0.9.28,<0.10.0)",
    "pandas (>=2.2.3)",
    "openpyxl (>=3.1.5)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0)",
    "black (>=25.11.0)",
    "poetry-core (>=2.2.1,<3.0.0)",
    "sqlacodegen (>=3.2.0,<4.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]
